refactor: Registration degree 필드 Float 타입으로 변경

+++ src/main/java/com/onedrinktoday/backend/domain/registration/dto/RegistrationRequest.java
@@ -17,7 +17,7 @@ public class RegistrationRequest {
   private Long regionId;
   private String drinkName;
   private DrinkType type;
-  private Integer degree;
+  private Float degree;
   private Integer sweetness;
   private Integer cost;
   private String description;
